backend/requirements.txt
fastapi
uvicorn
python-dotenv
aiomysql
tortoise-orm
